music_shop/pyproject.toml
[tool.poetry]
name = "file"
version = "0.1.0"
description = ""
authors = ["MargitStar <sovenokbymargo@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.6"
psycopg2 = "^2.9.1"
djangorestframework = "^3.12.4"
Pillow = "^8.3.1"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
